# Implement `set_names()` in C