--- a/src/Makevars
+++ b/src/Makevars
@@ -29,6 +29,7 @@ lib-files = \
 
 internal-files = \
         internal/arg.c \
+        internal/attr.c \
         internal/dots.c \
         internal/call.c \
         internal/env.c \