--- a/src/export/init.c
+++ b/src/export/init.c
@@ -131,7 +131,7 @@ extern sexp* rlang_find_var(sexp*, sexp*);
 extern sexp* rlang_env_bind_list(sexp*, sexp*, sexp*);
 extern sexp* rlang_glue_is_there();
 extern sexp* rlang_linked_version();
-extern sexp* rlang_names2(sexp*);
+extern sexp* rlang_names2(sexp*, sexp*);
 extern sexp* rlang_set_names(sexp*, sexp*, sexp*, sexp*);
 
 // Library initialisation defined below
@@ -303,7 +303,7 @@ static const r_callable r_callables[] = {
   {"rlang_env_bind_list",               (r_fn_ptr) &rlang_env_bind_list, 3},
   {"rlang_glue_is_there",               (r_fn_ptr) &rlang_glue_is_there, 0},
   {"rlang_linked_version",              (r_fn_ptr) &rlang_linked_version, 0},
-  {"rlang_names2",                      (r_fn_ptr) &rlang_names2, 1},
+  {"rlang_names2",                      (r_fn_ptr) &rlang_names2, 2},
   {"rlang_set_names",                   (r_fn_ptr) &rlang_set_names, 4},
   {NULL, NULL, 0}
 };