--- a/src/export/init.c
+++ b/src/export/init.c
@@ -132,6 +132,7 @@ extern sexp* rlang_env_bind_list(sexp*, sexp*, sexp*);
 extern sexp* rlang_glue_is_there();
 extern sexp* rlang_linked_version();
 extern sexp* rlang_names2(sexp*);
+extern sexp* rlang_set_names(sexp*, sexp*, sexp*, sexp*);
 
 // Library initialisation defined below
 sexp* rlang_library_load(sexp*);
@@ -303,6 +304,7 @@ static const r_callable r_callables[] = {
   {"rlang_glue_is_there",               (r_fn_ptr) &rlang_glue_is_there, 0},
   {"rlang_linked_version",              (r_fn_ptr) &rlang_linked_version, 0},
   {"rlang_names2",                      (r_fn_ptr) &rlang_names2, 1},
+  {"rlang_set_names",                   (r_fn_ptr) &rlang_set_names, 4},
   {NULL, NULL, 0}
 };
 
@@ -311,7 +313,6 @@ extern sexp* rlang_is_missing(sexp*, sexp*, sexp*, sexp*);
 extern sexp* rlang_call2_external(sexp*, sexp*, sexp*, sexp*);
 extern sexp* rlang_ext2_dots_values(sexp*, sexp*, sexp*, sexp*);
 extern sexp* rlang_exec(sexp*, sexp*, sexp*, sexp*);
-extern sexp* rlang_set_names(sexp*, sexp*, sexp*, sexp*);
 
 
 static const r_external externals[] = {
@@ -319,7 +320,6 @@ static const r_external externals[] = {
   {"rlang_call2_external",              (r_fn_ptr) &rlang_call2_external, 2},
   {"rlang_ext2_dots_values",            (r_fn_ptr) &rlang_ext2_dots_values, 6},
   {"rlang_exec",                        (r_fn_ptr) &rlang_exec, 2},
-  {"rlang_set_names",                   (r_fn_ptr) &rlang_set_names, 3},
   {NULL, NULL, 0}
 };