--- a/src/internal/internal.c
+++ b/src/internal/internal.c
@@ -17,12 +17,14 @@ void rlang_init_utils();
 void rlang_init_dots(sexp* ns);
 void rlang_init_expr_interp();
 void rlang_init_eval_tidy();
+void rlang_init_attr(sexp* ns);
 
 void rlang_init_internal(sexp* ns) {
   rlang_init_utils();
   rlang_init_dots(ns);
   rlang_init_expr_interp();
   rlang_init_eval_tidy();
+  rlang_init_attr(ns);
 
   rlang_zap = rlang_ns_get("zap!");