--- a/src/lib/formula.c
+++ b/src/lib/formula.c
@@ -31,6 +31,14 @@ bool r_f_has_env(sexp* f) {
   return r_is_environment(r_f_env(f));
 }
 
+bool r_is_formula(sexp* x, int scoped, int lhs) {
+  if (r_is_formulaish(x, scoped, lhs)) {
+    return r_node_car(x) == r_tilde_sym;
+  } else {
+    return false;
+  }
+}
+
 bool r_is_formulaish(sexp* x, int scoped, int lhs) {
   static const char* formulaish_names[2] = { "~", ":=" };