… NEWS.md
… src/lib/eval.c
… src/lib/fn.h
… src/lib/formula.h
… src/lib/replace-na.h
… src/lib/rlang.h
… src/lib/sym.c
… src/lib/sym.h
… src/lib/vec-chr.c
… src/lib/vec-chr.h