To do this, I've had to implement a few other things in C:

- `is_function()`

   - While I was there, I went ahead and did `is_closure()`, `is_primitive()`, `is_primitive_eager()` and `is_primitive_lazy()` too

- `is_formula()`

- `names2()`

I've created a new file, `internal/attr.c` to house the C implementations of `set_names()` and `names2()`.

For `names2()`, we fall back to the R level `names()` when `x` is an object to allow it to dispatch. We also handle pairlists and language objects specially, like `getAttrib()` does.

For `set_names()`, we always eventually set the names with `names<-()` rather than with `r_poke_names()`. The reason for this is that `do_namesgets()` has the ability to shallow copy _attributes_ using an ALTREP wrapper with `R_shallow_duplicate_attr()`. This isn't exposed to us yet from C, but calling `names<-` and getting a wrapper object is way faster than calling `r_maybe_duplicate()`, which will duplicate `x`.

I have allowed the new `r_as_function()` to call out to `rlang::as_function()` for simplicity. I have seen that you worked on a partial implementation of a C `as_function()` in vctrs, but the quosure handling here looks tricky to me. Feel free to take over and implement that if you want 😬 

```r
library(rlang)

x <- 1:50000 + 0L
a <- rep("x", 50000L)
```

```r
bench::mark(set_names(x, a), iterations = 100000)

# master
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 set_names(x, a)    6.7µs   9.32µs   105179.     121KB     72.6

# this PR
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 set_names(x, a)   1.03µs    1.5µs   649682.    7.81KB     110.
```

```r
bench::mark(is_function(1))

# master
#> # A tibble: 1 x 6
#>   expression          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 is_function(1)   1.81µs   2.14µs   452624.      18KB     90.5

# this PR
#> # A tibble: 1 x 6
#>   expression          min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>     <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 is_function(1)    264ns    354ns  2491092.    3.86KB        0
```