@DavisVaughan 2020-02-27
I'm not sure I understand the errors. I seem to be able to reproduce the one in `test-fn.R` line 273 if I run that test locally (even on master). If I run a full `test()` it doesn't show up, but if I run it manually it does. Odd